wininst: include the tox sdist by default

+++ wininst.py
@@ -1,5 +1,6 @@
 # Build an installer for Windows
 
+import argparse
 import os
 import shutil
 import sys
@@ -19,16 +20,39 @@ def test_platform():
         raise SystemExit('Running on wrong platform/interpreter')
 
 
-def install_package(distribution_name):
-    pip.main(['install', '--target={}'.format(PKGS_DIR), distribution_name])
+def find_tox_sdist():
+    tox_dist_dir = os.path.join('.tox', 'dist')
+    sdist, = os.listdir(tox_dist_dir)
+    assert sdist.endswith('.zip') or sdist.endswith('.tar.gz')
+    return os.path.join(tox_dist_dir, sdist)
 
 
+def pip_install(*distributions):
+    rc = pip.main(['install', '--target={}'.format(PKGS_DIR), *distributions])
+    if rc != 0:
+        raise SystemExit('pip install failed')
+
+
+parser = argparse.ArgumentParser(description='Create a Windows installer.')
+parser.add_argument('distribution', type=str, nargs='?',
+                    help='requirement specifier or distribution archive for '
+                         'the rinohtype version to include in the installer')
+parser.add_argument('-t', '--use-tox-sdist', action='store_true',
+                    help='install the tox-built distribution')
 
 if __name__ == '__main__':
+    args = parser.parse_args()
     test_platform()
-    shutil.rmtree('build')
-    shutil.rmtree(PKGS_DIR)
+    if os.path.exists('build'):
+        shutil.rmtree('build')
+    if os.path.exists(PKGS_DIR):
+        shutil.rmtree(PKGS_DIR)
     os.mkdir(PKGS_DIR)
-    install_package('rinohtype')
-    install_package('pygments')
+    if args.use_tox_sdist:
+        assert args.distribution is None
+        rinohtype_dist = find_tox_sdist()
+    else:
+        assert args.distribution is not None
+        rinohtype_dist = args.distribution
+    pip_install('pygments', rinohtype_dist)
     pynsist(['wininst.cfg'])
